Separate modules for DataFramePackage and VariationGenerator
