+++ oemof_b3/model/variations.py
@@ -1,0 +1,31 @@
+import copy
+import os
+
+
+class VariationGenerator:
+
+    def __init__(self, datapackage):
+
+        self.base_datapackage = datapackage
+
+    def create_variations(self, variations, destination):
+
+        for id, changes in variations.iterrows():
+
+            dp = self.create_var(self.base_datapackage, changes)
+
+            variation_dir = os.path.join(destination, str(id))
+
+            dp.to_csv_dir(variation_dir)
+
+    def create_var(self, dp, changes):
+
+        _dp = copy.deepcopy(dp)
+
+        changes = changes.to_dict()
+
+        for (resource, var_name), var_value in changes.items():
+
+            _dp.data[resource].loc[:, var_name] = var_value
+
+        return _dp
